requirements.txt
flake8==3.9.2
tox==3.23.1
pre-commit==2.12.1
pytest==6.2.4
accessify==0.3.1
retry==0.9.2
mypy==0.812
pylint==2.8.2
appium-python-client==1.1.0
pytest-rerunfailures==9.1.1
pytest-timeout==1.4.2
pytest-timeouts==1.2.1